Using from now on q as exit key
* made this work for all levels
* moved all the leveltimeout delayed endlevel stuff into the end level controller of sandbox
* each level MUST have EndLevel() implemented, this led to errors before, since in the
end level base EndLevel() called the end screen of the sandbox
* end level base should now  yield an error message if this method is called in the base class
