+++ YouVsVirus/Assets/Scenes/Levelgethome/EndLevelControllerLevelgethome.cs
@@ -1,6 +1,7 @@
 ﻿using UnityEngine;
 using Components;
-using UnityEngine.SceneManagement;
+using UnityEngine.UI;
+using TMPro;
 
 /// <summary>
 /// The end level controller for level 1.
@@ -9,6 +10,12 @@ using UnityEngine.SceneManagement;
 /// </summary>
 public class EndLevelControllerLevelgethome : EndLevelControllerBase
 {
+    public TMP_Text failText = null; //  or make public and drag
+    public override void EndLevel()
+    {   
+        CanvasFail.SetActive(true);
+        failText.text = "You pressed the exit key.";
+    }
     public GameObject CreateHumans;
 
     /// <summary>
