+++ YouVsVirus/Assets/Scenes/Levelgethome/YouVsVirus_Levelgethome.unity
@@ -1988,7 +1988,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1293853907}
-  - component: {fileID: 1293853906}
   - component: {fileID: 1293853908}
   m_Layer: 0
   m_Name: EndLevelController
@@ -1997,18 +1996,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &1293853906
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1293853905}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1ae52b6270c0ad148a94f5770b491669, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!4 &1293853907
 Transform:
   m_ObjectHideFlags: 0
@@ -2037,11 +2024,10 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   CanvasFail: {fileID: 2046121199}
   CanvasSucc: {fileID: 236133598}
-  NPCPrefab: {fileID: 8900475433866948943, guid: ab9d2a99c32f65a43a4c99a9379ce533,
-    type: 3}
   LevelTimeout: 0
   EndConditionMetDelay: 3
   playerExposed: 0
+  failText: {fileID: 1959990384}
   CreateHumans: {fileID: 742637914}
 --- !u!1 &1342456105
 GameObject:
