+++ YouVsVirus/Assets/Scenes/leveldisco/EndLevelControllerLeveldisco.cs
@@ -10,7 +10,11 @@ using TMPro;
 public class EndLevelControllerLeveldisco : EndLevelControllerBase
 {
     public TMP_Text failText = null; //  or make public and drag
-   
+    public override void EndLevel()
+    {
+        CanvasFail.SetActive(true);
+        failText.text = "You pressed the exit key.";
+    }
     /// <summary>
     /// Failure due to friend exposed
     /// </summary>
