+++ YouVsVirus/ProjectSettings/EditorBuildSettings.asset
@@ -18,32 +18,32 @@ EditorBuildSettings:
     path: Assets/Scenes/levelsupermarket/YouVsVirus_Levelsupermarket.unity
     guid: ff7f43c3ae86d964eb3ee08526f9aa70
   - enabled: 0
-    path: Assets/Scenes/EndScreenLevelgethome.unity
-    guid: c17c6cec783f8154591a62e695e49656
-  - enabled: 1
-    path: Assets/Scenes/EndScreenLevelsupermarket.unity
-    guid: 2b462fdfffe3459488b2dccddf28a520
+    path: 
+    guid: 00000000000000000000000000000000
+  - enabled: 0
+    path: 
+    guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/leveldisco/StartScreenLeveldisco_1.unity
     guid: e4d3adc168e27f1418f5bbc623a30d4f
-  - enabled: 1
-    path: Assets/Scenes/LevelCollectMasks/EndScreenLevelcollectmasks.unity
-    guid: cc22b6d4347f05b46a480a402e78f1cb
-  - enabled: 1
-    path: Assets/Scenes/LevelCollectMasks/YouVsVirus_Levelcollectmasks.unity
-    guid: abb8b2cb0675081409b4ec60e49b237a
+  - enabled: 0
+    path: 
+    guid: 00000000000000000000000000000000
+  - enabled: 0
+    path: 
+    guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/leveldisco/StartScreenLeveldisco_2.unity
     guid: 3067c6d83b88f9d43a72e2e68ae1c3fb
   - enabled: 1
     path: Assets/Scenes/leveldisco/YouVsVirus_Leveldisco.unity
     guid: 6883f131ec1e3564bb7ad7ddb54eb8d9
-  - enabled: 1
-    path: Assets/Scenes/leveldisco/EndScreenLeveldisco_succ.unity
-    guid: cc193b554d6594b43913d51e2566f272
-  - enabled: 1
-    path: Assets/Scenes/leveldisco/EndScreenLeveldisco_fail.unity
-    guid: 834ac2d051797964a9109e0a32f59444
+  - enabled: 0
+    path: 
+    guid: 00000000000000000000000000000000
+  - enabled: 0
+    path: 
+    guid: 00000000000000000000000000000000
   - enabled: 1
     path: Assets/Scenes/Sandbox/YouVsVirus_Sandbox.unity
     guid: daf57548ff731d149b1cb2d5ba551c3b
